requirements.txt
Flask>=2.0
Flask-Login
Flask-SocketIO
Flask-SQLAlchemy
PyMySQL
eventlet
redis
pillow
pdf2image
reportlab
pdfrw
WeasyPrint
twilio